refactor: populate areaid on hover card click
